pyproject.toml
[tool.poetry]
name = "emo-dev-mode-example"
version = "0.1.0"
description = ""
authors = ["yoha-o <ooishi408@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
emo-platform-api-sdk = "^1.0.3"
python-dotenv = "^1.0.1"
pyngrok = "^7.2.0"
fastapi = "^0.115.2"
uvicorn = "^0.31.1"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
